+++ src/overlaybd/otel/tracer_common.h
@@ -3,107 +3,53 @@
 
 #pragma once
 
-#include "opentelemetry/context/propagation/global_propagator.h"
-#include "opentelemetry/context/propagation/text_map_propagator.h"
-#include "opentelemetry/exporters/ostream/span_exporter_factory.h"
+#include "opentelemetry/context/propagation/text_map_carrier.h"
 #include "opentelemetry/nostd/shared_ptr.h"
-#include "opentelemetry/sdk/trace/exporter.h"
-#include "opentelemetry/sdk/trace/processor.h"
-#include "opentelemetry/sdk/trace/provider.h"
-#include "opentelemetry/sdk/trace/simple_processor_factory.h"
-#include "opentelemetry/sdk/trace/tracer_context.h"
-#include "opentelemetry/sdk/trace/tracer_context_factory.h"
-#include "opentelemetry/sdk/trace/tracer_provider_factory.h"
-#include "opentelemetry/trace/propagation/http_trace_context.h"
-#include "opentelemetry/trace/provider.h"
+#include "opentelemetry/trace/tracer.h"
 
 #include <grpcpp/grpcpp.h>
-#include <cstring>
-#include <iostream>
-#include <vector>
 
 using grpc::ClientContext;
 using grpc::ServerContext;
 
-namespace
-{
+namespace overlaybd {
+namespace otel {
+
 class GrpcClientCarrier : public opentelemetry::context::propagation::TextMapCarrier
 {
 public:
-  GrpcClientCarrier(ClientContext *context) : context_(context) {}
+  GrpcClientCarrier(ClientContext *context);
   GrpcClientCarrier() = default;
-  virtual opentelemetry::nostd::string_view Get(
-      opentelemetry::nostd::string_view /* key */) const noexcept override
-  {
-    return "";
-  }
+  
+  opentelemetry::nostd::string_view Get(
+      opentelemetry::nostd::string_view key) const noexcept override;
+  
+  void Set(opentelemetry::nostd::string_view key,
+           opentelemetry::nostd::string_view value) noexcept override;
 
-  virtual void Set(opentelemetry::nostd::string_view key,
-                   opentelemetry::nostd::string_view value) noexcept override
-  {
-    std::cout << " Client ::: Adding " << key << " " << value << "\n";
-    context_->AddMetadata(std::string(key), std::string(value));
-  }
-
+private:
   ClientContext *context_ = nullptr;
 };
 
 class GrpcServerCarrier : public opentelemetry::context::propagation::TextMapCarrier
 {
 public:
-  GrpcServerCarrier(ServerContext *context) : context_(context) {}
+  GrpcServerCarrier(ServerContext *context);
   GrpcServerCarrier() = default;
-  virtual opentelemetry::nostd::string_view Get(
-      opentelemetry::nostd::string_view key) const noexcept override
-  {
-    auto it = context_->client_metadata().find({key.data(), key.size()});
-    if (it != context_->client_metadata().end())
-    {
-      return opentelemetry::nostd::string_view(it->second.data(), it->second.size());
-    }
-    return "";
-  }
+  
+  opentelemetry::nostd::string_view Get(
+      opentelemetry::nostd::string_view key) const noexcept override;
+  
+  void Set(opentelemetry::nostd::string_view key,
+           opentelemetry::nostd::string_view value) noexcept override;
 
-  virtual void Set(opentelemetry::nostd::string_view /* key */,
-                   opentelemetry::nostd::string_view /* value */) noexcept override
-  {
-    // Not required for server
-  }
-
+private:
   ServerContext *context_ = nullptr;
 };
 
-void InitTracer()
-{
-  auto exporter = opentelemetry::exporter::trace::OStreamSpanExporterFactory::Create();
-  auto processor =
-      opentelemetry::sdk::trace::SimpleSpanProcessorFactory::Create(std::move(exporter));
-  std::vector<std::unique_ptr<opentelemetry::sdk::trace::SpanProcessor>> processors;
-  processors.push_back(std::move(processor));
-  // Default is an always-on sampler.
-  std::unique_ptr<opentelemetry::sdk::trace::TracerContext> context =
-      opentelemetry::sdk::trace::TracerContextFactory::Create(std::move(processors));
-  std::shared_ptr<opentelemetry::trace::TracerProvider> provider =
-      opentelemetry::sdk::trace::TracerProviderFactory::Create(std::move(context));
-  // Set the global trace provider
-  opentelemetry::sdk::trace::Provider::SetTracerProvider(provider);
+void InitTracer();
+void CleanupTracer();
+opentelemetry::nostd::shared_ptr<opentelemetry::trace::Tracer> get_tracer(std::string tracer_name);
 
-  // set global propagator
-  opentelemetry::context::propagation::GlobalTextMapPropagator::SetGlobalPropagator(
-      opentelemetry::nostd::shared_ptr<opentelemetry::context::propagation::TextMapPropagator>(
-          new opentelemetry::trace::propagation::HttpTraceContext()));
-}
-
-void CleanupTracer()
-{
-  std::shared_ptr<opentelemetry::trace::TracerProvider> none;
-  opentelemetry::sdk::trace::Provider::SetTracerProvider(none);
-}
-
-opentelemetry::nostd::shared_ptr<opentelemetry::trace::Tracer> get_tracer(std::string tracer_name)
-{
-  auto provider = opentelemetry::trace::Provider::GetTracerProvider();
-  return provider->GetTracer(tracer_name);
-}
-
-}  // namespace
+}  // namespace otel
+}  // namespace overlaybd
